+++ src/components/common/Routing.tsx
@@ -3,9 +3,9 @@ import BankBottomSheet from '@components/molecules/bottomSheet/BankBottomSheet';
 import LoginBottomSheet from '@components/molecules/bottomSheet/LoginBottomSheet';
 import ProgressEditAccountTemplate from '@components/templates/account/ProgressEditAccountTemplate';
 import ProgressRegisterAccountTemplate from '@components/templates/account/ProgressRegisterAccountTemplate';
-import PostWithdrawalScreen from '@components/templates/Withdrawal/PostWithdrawalScreen';
-import PreWithdrawalScreen from '@components/templates/Withdrawal/PreWithdrawalScreen';
-import ProgressWithdrawalScreen from '@components/templates/Withdrawal/ProgressWithdrawalScreen';
+import PostWithdrawalScreen from '@components/templates/withdrawal/PostWithdrawalScreen';
+import PreWithdrawalScreen from '@components/templates/withdrawal/PreWithdrawalScreen';
+import ProgressWithdrawalScreen from '@components/templates/withdrawal/ProgressWithdrawalScreen';
 import EnterAccountInfo from '@pages/account/EnterAccountInfo';
 import ReadAccountInfo from '@pages/account/ReadAccountInfo';
 import SaveAccountInfo from '@pages/account/SaveAccountInfo';
